Melhoria de layout na seção Informações do Pedido

+++ backend/app/models.py
@@ -161,7 +161,7 @@ class OrderItem(db.Model):
     )
 
     quantity = db.Column(
-        db.Integer,
+        db.Float,
         nullable=False,
         default=1
     )
